Increase static attr. timeout for perf. analysis
Reason: The former timeout in the order of seconds was an additional
benefit for very fast algorithms, as they were able to skip
re-evaluation of the attribute completely. As no common minimum time for
all algorithms can be found, we "removed" the limit by increasing it to
a couple of minutes.